chahua: 重构 renderer.js（七）—— 抽出 modals 模块
添加茶客 / 新建房间 / 新建任务 / 导入 persona 四个 modal 的装配抽到 modals.js：
模块自取所需 DOM 节点（仅此处用），renderer 经少量回调（isConnected / send /
setStatus / getGuests / getPersonas / pickFolder）+ openModal / closeModal 注入。
全 modal 的「点 backdrop / × / ESC 关闭」通用兜底也一并迁入。

renderer 的 ~18 个 modal 专属 DOM 常量随之删除，persona.js / buildOwnerOptionData
/ DEFAULT_PERMISSION 导入清掉。openModal / closeModal 留在 renderer（modals 与
decision_support 共用同一对）。

renderer.js 986 → 847 行。自原始 1671 累计 −824（−49%），god file 拆成 8 个
单一职责模块（message_filter / commands / connection / room_list / turn_state /
envelope_router / sidebar / modals）。行为零变化，9 个模块 node --check 通过。
